+++ src/SequenceItem.cpp
@@ -1,15 +1,15 @@
 #include "SequenceItem.h"
 
-SequenceItem::SequenceItem(int period) {
+SequenceItem::SequenceItem(long period) {
     this->period = period;
 }
 
-SequenceItem::SequenceItem(int period, bool active) {
+SequenceItem::SequenceItem(long period, bool active) {
     this->period = period;
     this->active = active;
 }
 
-int SequenceItem::getPeriod() {
+long SequenceItem::getPeriod() {
     return period;
 }
